feat: added Sidebar and moved NavbarItem to components directory
