+++ src/lib/components/nav-item.svelte
